requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
google-generativeai
python-dotenv
